// src/types/PersonalPage.ts
export interface HeroData {
    imageUrl: string;
    altText: string,
    title: string;
    buttonText: string;    
    buttonLink: string;
  }
  
  export interface HighlightData {
    link: string;
    iconSrc: string;
    altText: string;
    title: string;
    description: string;
  }

  export interface BioData {
    title: string;
    about: string;
  }
  

  